web/ssh/server: finish the C-ABI sweep — every integer is i64 at the boundary
Follow-up to the `server::listen` fix, which found the class but not all of it.  A
mechanical audit of `#native` declarations against their Rust exports across every lib
repo turned up 17 more mismatches here.

loft emits the extern from the DECLARATION (`integer` -> `i64`, PACKAGES.md) and calls it
directly, so a narrower Rust signature reads undefined register bits.  Six were RETURNS,
the dangerous half, and two of those defeated a live guard in loft code:

  n_ws_connect      -1 on failure; web.loft does `if id < 0 { return null; }`
                    -> on --native the check never fired and a failed connect handed
                       back a WsHandler with a garbage id
  n_ws_group_poll   -1 means "nothing ready"; web.loft does `if h < 0 { … }`
                    -> "idle" read as a valid handle 4294967295
  n_byte_at         -1 on out-of-range, consumed directly for byte assembly
  n_http_do         status code / negative transport error
  n_ws_opcode       `-> u8` (missed by the first pass, which only looked for i32/u32):
  n_ws_client_opcode  returning u8 sets only `al`, so loft read 56 undefined bits

The other 11 were PARAMS — lower risk (the caller writes 64 bits, the callee reads the
low 32, which is right for in-range values on x86-64) but off-contract just the same.

Widened every one at the boundary and shadowed straight back to the working width, so no
function body changed.  `server` also picks up the `n_ws_opcode` return the first pass
missed.

Verified: all three crates build; the server, web (byte_at + pack) and package tests pass
on BOTH backends.  The rule now has a permanent home in the loft repo —
`tests/native_abi_contract.rs` checks it statically, and `native_scalar_pkg` proves it
behaviourally with a -1 sentinel.

+++ server/native/src/lib.rs
@@ -494,8 +494,12 @@ pub extern "C" fn n_ws_message() -> LoftStr {
 /// Get the last received WebSocket opcode (1=text, 2=binary, 8=close, 9=ping, 10=pong).
 #[loft_native]
 #[unsafe(no_mangle)]
-pub extern "C" fn n_ws_opcode() -> u8 {
-    WS_LAST_OPCODE.with(|o| *o.borrow())
+// `i64` on the C boundary: the loft declaration types this `integer`, and loft emits
+// the extern from the DECLARATION, so a narrower Rust return leaves the upper half of
+// the register undefined and loft reads garbage — visibly, for negatives, on --native
+// only.  Narrow inside the body instead.
+pub extern "C" fn n_ws_opcode() -> i64 {
+    WS_LAST_OPCODE.with(|o| i64::from(*o.borrow()))
 }
 
 /// Send a text WebSocket message.
